requirements.txt
tensorflow_addons
tensorflow_io
tensorflow-model-optimization
click
scikit-learn
librosa
pandas